mobile application backend/requirements.txt
Flask
requests
python-dotenv
geopy
APScheduler  # Add the scheduler library